examples/python/openai_sdk_1.x/requirements.txt
openai>=1.3.3, <2.0.0
python_dotenv>=1.0.0, <2.0.0
azure-search-documents
httpx
tenacity
termcolor
black
ruff
